port launch files
the port is "done". in quotes because there's still a shit load of
warnings (missing textures that are not even in the repo, the fuck?) and
somehow some race conditions (like needing the pedestrian to launch
after the car. won't bother fixing yet). i will consider this done and
fix stuff later, we'll have a better idea of what needs fixing as we go.